requirements.txt
numpy
slackclient
python-dotenv
scikit-learn
pytesseract
Pillow
tensorflow
tensorflow-hub
tensorflow-text